+++ cmd/dep/hash_in.go
@@ -8,8 +8,8 @@ import (
 	"flag"
 
 	"github.com/golang/dep"
-	"github.com/golang/dep/internal/gps"
-	"github.com/golang/dep/internal/gps/pkgtree"
+	"github.com/golang/dep/gps"
+	"github.com/golang/dep/gps/pkgtree"
 	"github.com/pkg/errors"
 )