+++ test_project_context_test.go
@@ -7,7 +7,7 @@ package dep
 import (
 	"path/filepath"
 
-	"github.com/golang/dep/internal/gps"
+	"github.com/golang/dep/gps"
 	"github.com/golang/dep/internal/test"
 	"github.com/pkg/errors"
 )