+++ CHANGELOG.md
@@ -8,6 +8,7 @@ BUG FIXES:
 IMPROVEMENTS:
 
 * Log as dependencies are pre-fetched during dep init ([#1176](https://github.com/golang/dep/pull/1176)).
+* Make the gps package importable ([#1349](https://github.com/golang/dep/pull/1349)).
 
 # v0.3.2